HistogramThresholding.jl Documentation

HistogramThresholding.jl Documentation

HistogramThresholding.find_threshold — Method.

t = find_threshold(Otsu(),  histogram, edges)

Under the assumption that the histogram is bimodal the threshold is set so that the resultant inter-class variance is maximal.

Arguments

The function arguments are described in more detail below.

histogram

An AbstractArray storing the frequency distribution.

edges

An AbstractRange specifying how the intervals for the frequency distribution are divided.

Reference

Nobuyuki Otsu (1979). "A threshold selection method from gray-level histograms". IEEE Trans. Sys., Man., Cyber. 9 (1): 62–66. doi:10.1109/TSMC.1979.4310076

source